refactor: move url and env utils to shared
