+++ app/shared/utils/env-utils.ts
@@ -1,10 +1,10 @@
 
-function getEnv(identifier: string): string {
+function getPublicEnv(identifier: string): string {
   return Meteor.settings.public[identifier] ?? ""
 }
 
-export function getEnvOrWarn(identifier: string): string {
-  const envValue = getEnv(identifier)
+export function getPublicEnvOrWarn(identifier: string): string {
+  const envValue = getPublicEnv(identifier)
 
   if (!envValue) {
     console.warn('Warning: Environment variable (%s) is not defined.', identifier)
